selenium_behave_framework/requirements.txt
# Selenium Behave POM Framework Requirements

# Core Testing
selenium>=4.15.0
behave>=1.2.6
webdriver-manager>=4.0.1

# Excel/Data handling
openpyxl>=3.1.2
pandas>=2.0.0

# Reporting
allure-behave>=2.13.2
behave-html-formatter>=0.9.10

# Utilities
python-dotenv>=1.0.0
colorlog>=6.8.0
Pillow>=10.2.0

# Optional: For parallel execution
# behave-parallel>=0.0.8

# Optional: API testing integration
requests>=2.31.0

# Development/Testing
pytest>=7.4.3
pytest-selenium>=4.0.2

# Type hints and code quality
# mypy>=1.8.0
# flake8>=7.0.0
# black>=24.1.0
